requirements.txt
pandas
numpy
statsmodels
yfinance
matplotlib
scipy
pyyaml
